Add missing int and remove all_money temporarily

+++ backend/exchange/ftx.py
@@ -143,15 +143,15 @@ class FtxController(exchange.base.ExchangeBase):
         if descriptor.limit > 0:
             request_url += "&limit=" + str(descriptor.limit)
 
-        if descriptor.start_date is not None:
-            request_url += "&start_time=" + \
-                str((descriptor.start_date
-                     - datetime.datetime(1970, 1, 1)).total_seconds() * 1000)
+            if descriptor.start_date is not None:
+                request_url += "&start_time=" + \
+                    str(int((descriptor.start_date
+                             - datetime.datetime(1970, 1, 1)).total_seconds() * 1000))
 
-        if descriptor.end_date is not None:
-            request_url += "&end_time=" + \
-                str((descriptor.end
-                     - datetime.datetime(1970, 1, 1)).total_seconds() * 1000)
+            if descriptor.end_date is not None:
+                request_url += "&end_time=" + \
+                    str(int((descriptor.end
+                             - datetime.datetime(1970, 1, 1)).total_seconds() * 1000))
 
         logging.debug("FTX price history request: %s", request_url)
         response = requests.get(request_url)
