+++ backend/exchange/ftx_mock.py
@@ -52,13 +52,13 @@ class FtxMock(exchange.mock_base.MockBase):
 
             if descriptor.start_date is not None:
                 request_url += "&start_time=" + \
-                    str((descriptor.start_date
-                         - datetime.datetime(1970, 1, 1)).total_seconds() * 1000)
+                    str(int((descriptor.start_date
+                             - datetime.datetime(1970, 1, 1)).total_seconds() * 1000))
 
             if descriptor.end_date is not None:
                 request_url += "&end_time=" + \
-                    str((descriptor.end
-                         - datetime.datetime(1970, 1, 1)).total_seconds() * 1000)
+                    str(int((descriptor.end
+                             - datetime.datetime(1970, 1, 1)).total_seconds() * 1000))
 
             logging.debug("FTX price history request: %s", request_url)
             response = requests.get(request_url)
